Move remaining functions to dedicated submodules
