--- docs/api/autodoc/spglib.spglib.rst
@@ -1,8 +1,0 @@
-spglib.spglib module
-====================
-
-.. automodule:: spglib.spglib
-   :members:
-   :private-members:
-   :show-inheritance:
-   :undoc-members:
